Jesse's models

+++ hw1/submission/models.py
@@ -95,13 +95,12 @@ class LogisticRegression(nn.Module):
                                           mode='sum')
         self.embeddings.weight = nn.Parameter(torch.eye(len(TEXT.vocab)),
                                               requires_grad=False)
-        # Linear layer
         self.linear = nn.Linear(len(TEXT.vocab), len(LABEL.vocab))
-        
+    
     # Here bow is [len-of-sentence, N] -- it is an integer matrix
     def forward(self, bow):
         bow_features = self.embeddings(bow)
-        return F.log_softmax(self.linear(bow_features), dim=1)
+        return F.log_softmax(self.linear(bow_features.float()), dim=1)
     
 class CBOW(nn.Module):
     def __init__(self, TEXT, LABEL):
